Fix compilation errors
Fixes: https://github.com/me-no-dev/ESPAsyncTCP/issues/57
Fixes: https://github.com/me-no-dev/ESPAsyncTCP/pull/56
Closes: https://github.com/me-no-dev/ESPAsyncTCP/pull/56

+++ src/ESPAsyncTCP.cpp
@@ -281,7 +281,7 @@ size_t AsyncClient::ack(size_t len){
 
 // Private Callbacks
 
-int8_t AsyncClient::_connected(void* pcb, int8_t err){
+long AsyncClient::_connected(void* pcb, long err){
   _pcb = reinterpret_cast<tcp_pcb*>(pcb);
   if(_pcb){
     _pcb_busy = false;
@@ -334,7 +334,7 @@ int8_t AsyncClient::_close(){
   return err;
 }
 
-void AsyncClient::_error(int8_t err) {
+void AsyncClient::_error(long err) {
   if(_pcb){
 #if ASYNC_TCP_SSL_ENABLED
     if(_pcb_secure){
@@ -361,7 +361,7 @@ void AsyncClient::_ssl_error(int8_t err){
 }
 #endif
 
-int8_t AsyncClient::_sent(tcp_pcb* pcb, uint16_t len) {
+long AsyncClient::_sent(tcp_pcb* pcb, uint16_t len) {
   _rx_last_packet = millis();
   ASYNC_TCP_DEBUG("_sent: %u\n", len);
   _tx_unacked_len -= len;
@@ -375,7 +375,7 @@ int8_t AsyncClient::_sent(tcp_pcb* pcb, uint16_t len) {
   return ERR_OK;
 }
 
-int8_t AsyncClient::_recv(tcp_pcb* pcb, pbuf* pb, int8_t err) {
+long AsyncClient::_recv(tcp_pcb* pcb, pbuf* pb, long err) {
   if(pb == NULL){
     ASYNC_TCP_DEBUG("_recv: pb == NULL! Closing... %d\n", err);
     return _close();
@@ -414,7 +414,7 @@ int8_t AsyncClient::_recv(tcp_pcb* pcb, pbuf* pb, int8_t err) {
   return ERR_OK;
 }
 
-int8_t AsyncClient::_poll(tcp_pcb* pcb){
+long AsyncClient::_poll(tcp_pcb* pcb){
   // Close requested
   if(_close_pcb){
     _close_pcb = false;
@@ -469,23 +469,23 @@ void AsyncClient::_s_dns_found(const char *name, ip_addr_t *ipaddr, void *arg){
   reinterpret_cast<AsyncClient*>(arg)->_dns_found(ipaddr);
 }
 
-int8_t AsyncClient::_s_poll(void *arg, struct tcp_pcb *tpcb) {
+long AsyncClient::_s_poll(void *arg, struct tcp_pcb *tpcb) {
   return reinterpret_cast<AsyncClient*>(arg)->_poll(tpcb);
 }
 
-int8_t AsyncClient::_s_recv(void *arg, struct tcp_pcb *tpcb, struct pbuf *pb, int8_t err) {
+long AsyncClient::_s_recv(void *arg, struct tcp_pcb *tpcb, struct pbuf *pb, long err) {
   return reinterpret_cast<AsyncClient*>(arg)->_recv(tpcb, pb, err);
 }
 
-void AsyncClient::_s_error(void *arg, int8_t err) {
+void AsyncClient::_s_error(void *arg, long err) {
   reinterpret_cast<AsyncClient*>(arg)->_error(err);
 }
 
-int8_t AsyncClient::_s_sent(void *arg, struct tcp_pcb *tpcb, uint16_t len) {
+long AsyncClient::_s_sent(void *arg, struct tcp_pcb *tpcb, uint16_t len) {
   return reinterpret_cast<AsyncClient*>(arg)->_sent(tpcb, len);
 }
 
-int8_t AsyncClient::_s_connected(void* arg, void* tpcb, int8_t err){
+long AsyncClient::_s_connected(void* arg, void* tpcb, long err){
     return reinterpret_cast<AsyncClient*>(arg)->_connected(tpcb, err);
 }
 
@@ -896,7 +896,7 @@ uint8_t AsyncServer::status(){
   return _pcb->state;
 }
 
-int8_t AsyncServer::_accept(tcp_pcb* pcb, int8_t err){
+long AsyncServer::_accept(tcp_pcb* pcb, long err){
   if(_connect_cb){
 #if ASYNC_TCP_SSL_ENABLED
     if (_noDelay || _ssl_ctx)
@@ -964,7 +964,7 @@ int8_t AsyncServer::_accept(tcp_pcb* pcb, int8_t err){
   return ERR_OK;
 }
 
-  int8_t AsyncServer::_s_accept(void *arg, tcp_pcb* pcb, int8_t err){
+  long AsyncServer::_s_accept(void *arg, tcp_pcb* pcb, long err){
     return reinterpret_cast<AsyncServer*>(arg)->_accept(pcb, err);
   }
